requirements.txt
python-telegram-bot==21.7
requests==2.32.3
python-dotenv==1.0.1
Pillow>=10.0.0
scipy>=1.11.0
numpy>=1.24.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
